+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Bank Select and Morph State/inputs.py
@@ -1,0 +1,71 @@
+##############################################################################################################################################
+# 
+# Definition of actions for switches
+#
+##############################################################################################################################################
+ 
+from pyswitch.hardware.devices.pa_midicaptain_nano_4 import *
+
+from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
+from pyswitch.clients.kemper.actions.bank_select import BANK_SELECT
+from pyswitch.clients.kemper.actions.morph import MORPH_DISPLAY
+
+from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
+
+##############################################################################################################################################
+
+# Defines the switch assignments
+Inputs = [
+
+    # Switch 1
+    {
+        "assignment": PA_MIDICAPTAIN_NANO_SWITCH_1,
+        "actions": [
+            BANK_SELECT(
+                bank = 1,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                display = DISPLAY_HEADER_1
+            ),
+            MORPH_DISPLAY()
+        ]
+    },
+
+    # Switch 2
+    {
+        "assignment": PA_MIDICAPTAIN_NANO_SWITCH_2,
+        "actions": [
+            BANK_SELECT(
+                bank = 2,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                display = DISPLAY_HEADER_2
+            ),
+            MORPH_DISPLAY()   
+        ]
+    },
+
+    # Switch A
+    {
+        "assignment": PA_MIDICAPTAIN_NANO_SWITCH_A,
+        "actions": [
+            RIG_SELECT(
+                rig = 1,
+                display = DISPLAY_FOOTER_1,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG                
+            ),
+            MORPH_DISPLAY()
+        ]
+    },
+    
+    # Switch B
+    {
+        "assignment": PA_MIDICAPTAIN_NANO_SWITCH_B,
+        "actions": [
+            RIG_SELECT(
+                rig = 2,
+                display = DISPLAY_FOOTER_2,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            ),
+            MORPH_DISPLAY()
+        ]
+    },
+]